train/requirements.txt
# conda install pytorch==1.8.0 torchvision==0.9.0 torchaudio==0.8.0 cudatoolkit=10.2 -c pytorch

opencv-python==4.2.0.34
psutil
h5py
matplotlib
scipy
scikit-image==0.16.2
tqdm
tensorboard
torchsummary
statistics
tables
pyyaml
yacs
pathlib
termcolor
nvidia_ml_py3
pickle5
trimesh
numba
numpy-quaternion
seaborn
sympy
icecream
pytorch-lightning
shapely
vtk
tensorboardX
timm
iopath
ipympl